+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Change: Drop compatibility for version of WP older than 4.9
 * Change: Update links to coffee2code.com to be HTTPS
 * Unit tests:
+    * New: Add tests for `admin_init()`
     * New: Add test for setting name
     * Change: Update tests for default hooks, removing a duplicate and adding 2 that were missing
     * Change: Use HTTPS for link to WP SVN repository in bin script for configuring unit tests (and delete commented-out code)